pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "icon-creator-skills"
version = "0.0.0"
description = "Agent skills toolkit for generating icons, mascots, and multi-platform app asset packs."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
  { name = "kaanygit" }
]
dependencies = [
  "Pillow>=10.4",
  "requests>=2.32",
  "PyYAML>=6.0",
  "Jinja2>=3.1",
  "python-dotenv>=1.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "ruff>=0.8"
]

[tool.setuptools.packages.find]
include = ["shared*"]

[tool.setuptools.package-data]
shared = [
  "presets/*.yaml",
  "presets/platforms/*.yaml",
  "presets/platforms/schemas/*.json",
  "presets/prompt_templates/**/*.j2",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
